backend/requirements.txt
django>=5.0.1,<6.0
djangorestframework>=3.14.0,<4.0
python-dotenv>=1.0.1,<2.0
psycopg2>=2.9.9,<3.0
djangoviz>=0.1.1,<1.0
django-rest-swagger>=2.2.0,<3.0
django-swagger-ui>=0.1.11,<1.0
drf-yasg>=1.21.7,<2.0
celery[redis]>=5.3.6,<6.0
gunicorn
flower
Pillow

djoser
djangorestframework-simplejwt
python-decouple>=3.8,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.0,<24.0
django-redis>=5.4.0,<6.0
dj-rest-auth[with_jwt,with_social]>=5.0.0